pyproject.toml
[project]
name = "chartbeacon"
version = "1.0.0"
description = "Personal technical indicators dashboard with Discord notifications"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "sqlalchemy>=1.4.36,<2.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.7.2",
    "python-dotenv>=1.0.1",
    "yfinance>=0.2.40",
    "pandas>=2.2.2",
    "psycopg2-binary>=2.9.9",
    "numpy==1.26.3",
    "requests>=2.32.3",
    "apache-airflow>=2.9.1",
    "setuptools>=68.0.0",
    "pandas-ta==0.3.14b0",
]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "black>=24.4.2",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "mypy>=1.10.0",
]
airflow = [
    "yfinance>=0.2.40",
    "pandas>=2.2.2",
    "pandas-ta==0.3.14b0",
    "apache-airflow>=2.9.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.uv]
managed = true
default-groups = ["dev"]

[tool.hatch.build.targets.wheel]
packages = ["api", "airflow"] 
